feat: complete frameless window with titlebar
